table setup script working
added airflow variables json file

added dag defs file

+++ src/defs/bq/personalization.py
@@ -97,7 +97,7 @@ SCHEMAS = {
         },
         {
             "name": "product_embedding",
-            "type": "FLOAT32",
+            "type": "FLOAT64",
             "mode": "REPEATED"
         },
         {
